Implement the GameDataStore
