+++ Assets/DataStore.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 4e7186636fe917241a59ce9393999119
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
