+++ Assets/DataStore/BaristaDialogue.cs
@@ -1,0 +1,44 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+[CreateAssetMenu(fileName = "BaristaDialogue", menuName = "Barista Dialogue")]
+public class BaristaDialogue : ScriptableObject
+{
+
+    #region - - - - - - Fields - - - - - -
+
+    [Header("Barista Dialogue")]
+    [Space]
+    public List<DialogueObject> dialogueObjects;
+
+    #endregion Fields
+
+}
+
+
+[System.Serializable]
+public class DialogueObject
+{
+
+    #region - - - - - - Fields - - - - - -
+
+    /// <summary>
+    /// The unique identifier for the dialogue object.
+    /// </summary>
+    public string identifier;
+
+    /// <summary>
+    /// The question literal for the dialogue.
+    /// </summary>
+    public string question;
+
+    /// <summary>
+    /// The optional response literals for additional dialogue.
+    /// </summary>
+    public List<string> response;
+
+    #endregion Fields
+
+}
+
