+++ Assets/DataStore/GameDataStore.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 8f204f6da2cd3404c914b4cc6ad454f6
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
